refactor: 테스트 파일 경로 이동
- 테스트 파일 컴포넌트 폴더 밑 __test__ 경로로 이동
- 주석 제거
