+++ client/src/components/test-example/__test__/NotFound.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import NotFound from './NotFound';
+import NotFound from '../NotFound';
 
 describe('<NotFound />', () => {
   it('renders header', () => {
